Updated RedBlackTree so it implements IEnumerable<T>. Updated logic to update note time and positions based on timing event changes to call into existing SMCommon logic, using the IEnumerable<T> implementation of RedBlackTree. Renamed MiscEventWidgetManager to MiscEventWidgetLayoutManager since it just manages layout. Made many EditorEvent fields private or protected with accessors to allow subclasses more control. Specifically this enables a hacky workaround to width computation which relies on ImGui. Misc events are now rendered in all scroll modes. Removed placeholder implicit operators on EditorEvents left over from last commit. Finished splitting EditorEvent subclasses out to their own files. Added interpolated scroll rate misc event widgets.

+++ Fumen/Converters/SMCommon.cs
@@ -527,7 +527,7 @@ namespace Fumen.Converters
 		{
 			foreach (var scrollRate in scrollRateEvents)
 			{
-				var scrollRateEvent = new ScrollRate((float)scrollRate.Value)
+				var scrollRateEvent = new ScrollRate(scrollRate.Value)
 				{
 					IntegerPosition = ConvertAbsoluteBeatToIntegerPosition(scrollRate.Key)
 				};
@@ -568,7 +568,7 @@ namespace Fumen.Converters
 					periodAsIntegerPosition = ConvertAbsoluteBeatToIntegerPosition(length);
 				}
 
-				var scrollRateEvent = new ScrollRateInterpolation((float)speed, periodAsIntegerPosition, periodAsTimeMicros, lengthIsTimeInSeconds)
+				var scrollRateEvent = new ScrollRateInterpolation(speed, periodAsIntegerPosition, periodAsTimeMicros, lengthIsTimeInSeconds)
 				{
 					IntegerPosition = ConvertAbsoluteBeatToIntegerPosition(scrollRate.Key)
 				};
